Add accounts_password_pam_unix_enabled to RHEL 10 CIS
Add rule accounts_password_pam_unix_enabled to RHEL 10 CIS
and update it for RHEL 10. The rule has originally been written
for Ubuntu, it needed modifications for RHEL 10 because the
configuration is little different.

The rule doesn't have remediation on RHEL 10 because the
remediation is covered by a different control in CIS and
users should use authselect to configure PAM modules.
Notice that the remediation code in the CIS requirement
5.3.1.5 is a read-only script.

Resolves: https://issues.redhat.com/browse/OPENSCAP-6125