Remove accounts_user_dot_no_world_writable_programs from CIS
We have reviewed requirement 7.2.9 in CIS RHEL 9 Benchmark
v2.0.0 and requirement 7.2.9 in CIS RHEL 10 Benchmark v1.0.1.
We think that the rule `accounts_user_dot_no_world_writable_programs`
doesn't need to be present in the CIS profiles. We haven't found
any text that would justify selecting this rule in CIS profiles.
This rule is problematic and dubious we should happily remove it.

Related to https://issues.redhat.com/browse/RHEL-135787