Shadow test scenario
Replace templated test scenario service_disabled.fail.sh by a
rule-specific test scenario of the same file name in rule
service_systemd-journal-upload_enabled. The reason is that the
systemd-journal-upload service fails to start if a remote logging URL
isn't configured on the system. The new test scenario configures this
URL and is very similar to already shadowing pass test scenario in
this rule.

Resolves: https://github.com/ComplianceAsCode/content/issues/14210